feat: Refactor chat widget into modular components and implement custom hooks for improved state management and functionality
