+++ components/chat-widget/hooks/index.ts
@@ -1,0 +1,5 @@
+export { useChatSuggestions } from './useChatSuggestions';
+export { useScrollVisibility } from './useScrollVisibility';
+export { useTypeWriter } from './useTypeWriter';
+export { useChatMessages } from './useChatMessages';
+export { useChatState } from './useChatState';
