+++ components/chat-widget/hooks/useChatMessages.ts
@@ -1,0 +1,99 @@
+'use client';
+
+import { useState, useRef, useEffect } from 'react';
+import { useTranslations } from 'next-intl';
+import { Message } from '../types';
+
+export const useChatMessages = () => {
+  const [messages, setMessages] = useState<Message[]>([]);
+  const [inputMessage, setInputMessage] = useState('');
+  const [isLoading, setIsLoading] = useState(false);
+  const messagesEndRef = useRef<HTMLDivElement>(null);
+  const t = useTranslations('chatWidget');
+
+  // Auto scroll to bottom when new messages arrive
+  useEffect(() => {
+    messagesEndRef.current?.scrollIntoView({ behavior: 'smooth' });
+  }, [messages]);
+
+  const sendMessage = async (message: string = inputMessage, onTypeWriter?: (text: string, callback?: () => void) => void) => {
+    if (!message.trim() || isLoading) return;
+
+    const userMessage: Message = {
+      role: 'user',
+      content: message.trim(),
+      timestamp: new Date(),
+    };
+
+    setMessages(prev => [...prev, userMessage]);
+    setInputMessage('');
+    setIsLoading(true);
+
+    try {
+      const response = await fetch('/api/chat', {
+        method: 'POST',
+        headers: {
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ message: message.trim() }),
+      });
+
+      const data = await response.json();
+
+      if (!response.ok) {
+        throw new Error(data.error || t('errorMessage'));
+      }
+
+      if (onTypeWriter) {
+        onTypeWriter(data.response, () => {
+          const assistantMessage: Message = {
+            role: 'assistant',
+            content: data.response,
+            timestamp: new Date(),
+          };
+          setMessages(prev => [...prev, assistantMessage]);
+        });
+      } else {
+        const assistantMessage: Message = {
+          role: 'assistant',
+          content: data.response,
+          timestamp: new Date(),
+        };
+        setMessages(prev => [...prev, assistantMessage]);
+      }
+    } catch (error) {
+      const errorMessage: Message = {
+        role: 'assistant',
+        content: error instanceof Error ? error.message : t('errorMessage'),
+        timestamp: new Date(),
+      };
+      setMessages(prev => [...prev, errorMessage]);
+    } finally {
+      setIsLoading(false);
+    }
+  };
+
+  const resetChat = () => {
+    setMessages([]);
+    setInputMessage('');
+    setIsLoading(false);
+  };
+
+  const handleKeyPress = (e: React.KeyboardEvent) => {
+    if (e.key === 'Enter' && !e.shiftKey) {
+      e.preventDefault();
+      sendMessage();
+    }
+  };
+
+  return {
+    messages,
+    inputMessage,
+    setInputMessage,
+    isLoading,
+    messagesEndRef,
+    sendMessage,
+    resetChat,
+    handleKeyPress,
+  };
+};
