+++ components/chat-widget/index.ts
@@ -1,0 +1,13 @@
+export { ChatHeader } from './ChatHeader';
+export { WelcomeMessage } from './WelcomeMessage';
+export { MessageBubble } from './MessageBubble';
+export { LoadingIndicator } from './LoadingIndicator';
+export { TypingIndicator } from './TypingIndicator';
+export { ChatInput } from './ChatInput';
+export { ChatWidgetBar } from './ChatWidgetBar';
+export { default as ChatWidget } from './ChatWidget';
+export { fetchSuggestions } from './api';
+export type { Message } from './types';
+
+// Export hooks
+export * from './hooks';
