+++ components/chat-widget/types.ts
@@ -1,0 +1,5 @@
+export interface Message {
+  role: 'user' | 'assistant';
+  content: string;
+  timestamp: Date;
+}
