+++ components/layout/Layout.tsx
@@ -1,6 +1,7 @@
 import Header from './Header';
 import Footer from './Footer';
 import ChatWidget from '../ChatWidget';
+import ClientProvider from '../ClientProvider';
 
 interface LayoutProps {
   children: React.ReactNode;
@@ -11,7 +12,9 @@ const Layout = ({ children }: LayoutProps) => {
     <div className="flex flex-col min-h-screen ">
       <Header />
       <main className="flex-grow container mx-auto px-6 py-8">{children}</main>
-      <ChatWidget />
+      <ClientProvider>
+        <ChatWidget />
+      </ClientProvider>
       <Footer />
     </div>
   );
